reproducing results with fat-fast

+++ examples/place_recognition/scriptRodaTodos.sh
@@ -1,26 +1,30 @@
 #!/bin/csh
 
-echo fatfast-100 em função de r
-foreach NEURON_MEMORY_INDEX_SIZE ( 1 5 10 15 )
+echo fatfast-100 em funcao de r
+foreach NEURON_MEMORY_INDEX_SIZE ( 1 5 10 15 20 25 30 )
+	set resultado = resultado_fatfastrand_100_r_${NEURON_MEMORY_INDEX_SIZE}.txt
+	if ! (-f "$resultado" ) then
+		echo fatfast_RAND_${NEURON_MEMORY_INDEX_SIZE}
 
-	echo fatfast_RAND_${NEURON_MEMORY_INDEX_SIZE}
+		rm -rf place_recognition.con
 
-	rm -rf place_recognition.con
+		sed -r "s/set NEURON_MEMORY_INDEX_SIZE(.*?)= (.*?);/set NEURON_MEMORY_INDEX_SIZE = $NEURON_MEMORY_INDEX_SIZE;/" place_recognition.con.source | sed -r "s/minchinton/minchinton_fat_fast_rand/" > place_recognition.con
 
-	sed -r "s/set NEURON_MEMORY_INDEX_SIZE(.*?)= (.*?);/set NEURON_MEMORY_INDEX_SIZE = $NEURON_MEMORY_INDEX_SIZE;/" place_recognition.con.source | sed -r "s/minchinton/minchinton_fat_fast_rand/" > place_recognition.con
+		make clean; make -f Makefile.no_interface VERBOSE=0
 
-	make clean; make -f Makefile.no_interface VERBOSE=0
+		mv place_recognition.con place_recognition_r_${NEURON_MEMORY_INDEX_SIZE}.con 
 
-	mv place_recognition.con place_recognition_r_${NEURON_MEMORY_INDEX_SIZE}.con 
-
-	./place_recognition place_recognition_original_100.cml > resultado_fatfastrand_100_r_${NEURON_MEMORY_INDEX_SIZE}.txt
+		./place_recognition place_recognition_original_100.cml > resultado_fatfastrand_100_r_${NEURON_MEMORY_INDEX_SIZE}.txt
+	endif
 end
 
 
 echo fatfast-zip-r-alto
-set NEURON_MEMORY_INDEX_SIZE = 15
+set NEURON_MEMORY_INDEX_SIZE = 30
 foreach CLUSTERING ( kmeans )
 	foreach COMPRESSION ( 001 045 075 )
+	set resultado = resultado_fatfastrand_${CLUSTERING}_${COMPRESSION}_r_${NEURON_MEMORY_INDEX_SIZE}.txt
+	if ! ( -f "$resultado" ) then
 
 		echo fatfast_${CLUSTERING}_${COMPRESSION}
 
@@ -33,6 +37,7 @@ foreach CLUSTERING ( kmeans )
 		mv place_recognition.con place_recognition_r_${NEURON_MEMORY_INDEX_SIZE}.con 
 
 		./place_recognition place_recognition_${CLUSTERING}_${COMPRESSION}.cml > resultado_fatfastrand_${CLUSTERING}_${COMPRESSION}_r_${NEURON_MEMORY_INDEX_SIZE}.txt
+	endif
 	end
 end
 
@@ -41,6 +46,8 @@ echo fatfast-100-r-baixo
 set NEURON_MEMORY_INDEX_SIZE = 1
 foreach CLUSTERING ( kmeans )
 	foreach COMPRESSION ( 001 045 075 )
+	set resultado = resultado_fatfastrand_${CLUSTERING}_${COMPRESSION}_r_${NEURON_MEMORY_INDEX_SIZE}.txt
+	if ! ( -f "$resultado" ) then
 
 		echo fatfast_${CLUSTERING}_${COMPRESSION}
 
@@ -53,6 +60,7 @@ foreach CLUSTERING ( kmeans )
 		mv place_recognition.con place_recognition_r_${NEURON_MEMORY_INDEX_SIZE}.con 
 
 		./place_recognition place_recognition_${CLUSTERING}_${COMPRESSION}.cml > resultado_fatfastrand_${CLUSTERING}_${COMPRESSION}_r_${NEURON_MEMORY_INDEX_SIZE}.txt
+	endif
 	end
 end
 
@@ -60,6 +68,8 @@ end
 echo vgram-zip
 foreach CLUSTERING ( kmeans )
 	foreach COMPRESSION ( 001 045 075 )
+	set resultado = resultado_${CLUSTERING}_${COMPRESSION}.txt
+	if ! ( -f "$resultado" ) then
 		 
 		echo vgram_${CLUSTERING}_${COMPRESSION}
 
@@ -68,6 +78,7 @@ foreach CLUSTERING ( kmeans )
 		make clean; make -f Makefile.no_interface VERBOSE=0
 
 		./place_recognition place_recognition_${CLUSTERING}_${COMPRESSION}.cml > resultado_${CLUSTERING}_${COMPRESSION}.txt
+	endif
 	end
 end
 
@@ -78,5 +89,7 @@ cp place_recognition.con.source place_recognition.con
 
 make clean; make -f Makefile.no_interface VERBOSE=0
 
-./place_recognition place_recognition_original_100.cml > resultado_original_100.txt
-
+set resultado = resultado_original_100.txt
+if ! ( -f "$resultado" ) then
+	./place_recognition place_recognition_original_100.cml > resultado_original_100.txt
+endif
